refactor(digitize): reorganise into packages and remove duplicate code
Restructure services/digitize from a flat module layout into focused
sub-packages, then remove the dead delegation wrappers that were left
behind after the reorganisation.

## Package reorganisation

Flat files → new locations (git-tracked as renames):

  cleanup.py        → pipeline/cleanup.py
  digitize_core.py  → pipeline/digitize.py
  ingest.py         → pipeline/ingest.py
  db_operations.py  → utils/db.py
  digitize_utils.py → utils/jobs.py        (partial — see below)
  docling_utils.py  → parsing/converter.py
  pdf_utils.py      → parsing/pdf.py
  docx_utils.py     → parsing/docx.py
  doc_utils.py      → processing/orchestrator.py + processing/{language,tables,text}.py

New files added as part of the reorganisation:
  api/v1/jobs.py, api/v1/documents.py  — endpoint routers extracted from app.py
  utils/storage.py                     — StorageManager: all file-system ops
  utils/recovery.py                    — zombie-job recovery on startup
  workers/concurrency.py               — BoundedSemaphore wrapper

## Duplicate/dead code removed

After the reorganisation utils/jobs.py retained three thin wrappers that
only delegated to StorageManager, creating a pointless 3-layer call chain:

  delete_document_files()    — dead: its only caller (api/v1/documents.py)
                               already called storage_manager.delete_document_content()
                               directly; the wrapper was never reached.

  cleanup_digitized_files()  — single-line passthrough to
                               storage_manager.delete_all_content();
                               had exactly one caller (bulk_delete_all_documents).

  bulk_delete_all_documents() — wrapped cleanup_digitized_files(), adding
                                only a constant and two log lines; had
                                exactly one caller (pipeline/cleanup.py).

Fix: remove all three functions and update pipeline/cleanup.py to call
storage_manager.delete_all_content() directly.

## Tests

- Remove monkeypatch of the deleted delete_document_files attribute from
  the test fixture in tests/test_digitize_app_endpoints.py.
- All 77 unit tests pass (PYTHONPATH=services .venv/bin/pytest -s).
