+++ services/digitize/api/v1/__init__.py
@@ -1,0 +1,1 @@
+# API v1 package
