+++ services/digitize/parsing/converter.py
@@ -1,3 +1,10 @@
+"""
+Docling conversion engine wrapper.
+
+Encapsulates all interaction with the Docling library:
+DocumentConverter setup, chunked conversion, format export.
+Moved from digitize/docling_utils.py as part of the parsing/ package reorganisation.
+"""
 import logging
 import shutil
 import tempfile
@@ -9,7 +16,7 @@ from typing import Optional
 from common.misc_utils import get_logger, DoclingConversionError
 from common.retry_utils import retry_on_transient_error
 from digitize.settings import settings
-from digitize.pdf_utils import get_document_page_count
+from digitize.parsing.pdf import get_document_page_count
 from digitize.models import OutputFormat
 
 # Docling document conversion libraries
@@ -92,8 +99,6 @@ def convert_doc(path: str | Path, cache_dir: Optional[Path] = None) -> DoclingDo
         return _convert_single_doc()
 
     # Process in chunks
-    # Calculate total chunks using ceiling division for the configured PDF chunk size.
-    # This ensures all pages are covered even if the last chunk is smaller.
     total_chunks = (total_pages + settings.digitize.doc_chunk_size - 1) // settings.digitize.doc_chunk_size
     logger.debug(
         f"Converting {path} document with {total_pages} pages in {total_chunks} "
@@ -109,7 +114,6 @@ def convert_doc(path: str | Path, cache_dir: Optional[Path] = None) -> DoclingDo
     chunk_cache_dir.mkdir(parents=True, exist_ok=True)
 
     try:
-        # Process document in chunks and save each chunk
         chunk_files = []
 
         for start_page in range(1, total_pages + 1, settings.digitize.doc_chunk_size):
@@ -120,16 +124,13 @@ def convert_doc(path: str | Path, cache_dir: Optional[Path] = None) -> DoclingDo
             chunk_file = convert_chunk(doc_converter, path, chunk_num, start_page, end_page, chunk_cache_dir)
             chunk_files.append(chunk_file)
 
-        # Load all chunk documents and concatenate
         docs = [DoclingDocument.load_from_json(filename=f) for f in chunk_files]
         concatenated_doc = DoclingDocument.concatenate(docs=docs)
 
         logger.debug(f"Successfully concatenated {path}'s {len(docs)} chunks into single document")
-        
         return concatenated_doc
 
     finally:
-        # Always cleanup cache directory
         try:
             shutil.rmtree(chunk_cache_dir)
             logger.debug(f"Cleaned up cache directory: {chunk_cache_dir}")
