+++ services/digitize/processing/__init__.py
@@ -1,0 +1,1 @@
+# processing package — doc_utils.py split into domain sub-modules
